requirements.txt
 
pillow
cryptography
numpy==1.24.4
opencv-python==4.5.5.64
scikit-learn==1.0.2
